pyproject.toml
[tool.poetry]
name = "stitch"
version = "0.1.0"
description = ""
authors = ["bnm <bnm@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
setuptools = "^62.1.0"
requests = "^2.28.1"
PyJWT = { extras = ["crypto"], version = "^2.4.0" }

[tool.poetry.group]
[tool.poetry.group.dev.dependencies]
bandit = "^1.7.4"
black = "^22.3.0"
flake8 = "^4.0.1"
flake8-bugbear = "^22.3.23"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.8.0"
flake8-isort = "^4.1.1"
flake8-pytest-style = "^1.6.0"
flake8-pytest = "^1.3"
pre-commit = "^2.18.1"
pytest = "^7.1.1"

[tool.poetry.group.example.dependencies]
Flask = "^2.2.2"
Flask-Cors = "^3.0.10"
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.bandit]
exclude_dirs = [".venv/", ".git/"]
assert_used = { skips = ["*/test_*.py", "*/test_*.py", "*/conftest.py"] }
